Rename SeedGenerator.js to SeedGeneratorAPI.js
